@@ -13,8 +13,8 @@ func TestSet(t *testing.T) {
 
 	// change all damages to negative
 	for i := 1; i <= ROUND; i++ {
-		filter := moper.NewD().Equal("damage", i)
-		update := moper.NewD().Set(moper.P{K: "damage", V: -i})
+		filter := moper.Query().Equal("damage", i)
+		update := moper.Query().Set(moper.P{K: "damage", V: -i})
 
 		result, err := mocom.UpdateMany[Hero](ctx, filter, update)
 		if err != nil {
@@ -30,8 +30,8 @@ func TestSet(t *testing.T) {
 
 	// change damages to positive
 	for i := 1; i <= ROUND; i++ {
-		filter := moper.NewD().Equal("damage", -i)
-		update := moper.NewD().Set(moper.P{K: "damage", V: i})
+		filter := moper.Query().Equal("damage", -i)
+		update := moper.Query().Set(moper.P{K: "damage", V: i})
 
 		result, err := mocom.UpdateMany[Hero](ctx, filter, update)
 		if err != nil {
@@ -49,9 +49,9 @@ func TestSet(t *testing.T) {
 func TestInc(t *testing.T) {
 	ctx := context.Background()
 	for i := ROUND; i >= 0; i-- {
-		filter := moper.NewD().Equal("damage", i)
+		filter := moper.Query().Equal("damage", i)
 
-		update := moper.NewD().Inc(moper.P{K: "damage", V: i})
+		update := moper.Query().Inc(moper.P{K: "damage", V: i})
 
 		result, err := mocom.UpdateMany[Hero](ctx, filter, update)
 		if err != nil {
@@ -66,8 +66,8 @@ func TestInc(t *testing.T) {
 	}
 
 	for i := 1; i <= ROUND; i++ {
-		filter := moper.NewD().Equal("damage", i*2)
-		update := moper.NewD().Inc(moper.P{
+		filter := moper.Query().Equal("damage", i*2)
+		update := moper.Query().Inc(moper.P{
 			K: "damage",
 			V: -i,
 		})
@@ -88,8 +88,8 @@ func TestInc(t *testing.T) {
 func TestPush(t *testing.T) {
 	ctx := context.Background()
 
-	filter := moper.NewD().Equal("damage", ROUND)
-	update := moper.NewD().Push(moper.P{K: "skillIds", V: 6})
+	filter := moper.Query().Equal("damage", ROUND)
+	update := moper.Query().Push(moper.P{K: "skillIds", V: 6})
 
 	result, err := mocom.UpdateMany[Hero](ctx, filter, update)
 	if err != nil {
@@ -97,7 +97,7 @@ func TestPush(t *testing.T) {
 		return
 	}
 
-	filter2 := moper.NewD().Equal("skillIds", []int{1, 2, 3, 4, 5, 6})
+	filter2 := moper.Query().Equal("skillIds", []int{1, 2, 3, 4, 5, 6})
 
 	if count, err := mocom.Count[Hero](ctx, filter2); err != nil {
 		t.Error("[TestPush]", err)
