@@ -2,7 +2,6 @@ package moper
 
 import "go.mongodb.org/mongo-driver/bson"
 
-func (d *D) Or(filters ...D) *D {
-	*d = append(*d, bson.E{Key: "$or", Value: filters})
-	return d
+func (d D) Or(filters ...D) D {
+	return append(d, bson.E{Key: "$or", Value: filters})
 }
